+++ pf4j/src/main/java/org/pf4j/ServiceProviderExtensionFinder.java
@@ -93,7 +93,7 @@ public class ServiceProviderExtensionFinder extends AbstractExtensionFinder {
             final Set<String> bucket = new HashSet<>();
 
             try {
-                Enumeration<URL> urls = findStorageResources(plugin.getPluginClassLoader(), EXTENSIONS_RESOURCE);
+                Enumeration<URL> urls = findStorageResources(plugin, EXTENSIONS_RESOURCE);
                 if (urls.hasMoreElements()) {
                     collectExtensions(urls, bucket);
                 } else {